Merged current backend and frontend code

+++ app/controllers/dream/index.js
@@ -1,15 +1,20 @@
+const mongoose = require('mongoose')
 const DreamServices = require('../../services').Dream
-const { Response } = require('../../utils/response')
+const { Response, response } = require('../../utils/response')
 
 module.exports = {
     getDreams: async (req, res) => {
         try {
             // Get request data
-            const pagination = req.query.pagination
+            const filter = req.query.search
+            const pagination = {
+                page: req.query.page,
+                limit: 10,
+            }
 
             // Get dreams from services
             const dreamServices = new DreamServices()
-            const dreams = await dreamServices.getDreams({ pagination })
+            const dreams = await dreamServices.getDreams(filter, pagination)
 
             // Return response
             res.send(new Response({ entity: { dreams } }))
@@ -21,35 +26,18 @@ module.exports = {
 
     createDream: async (req, res) => {
         try {
-            // FIXME: get user
             // Get request data
-            // Method 1:
-            // const { title, content, tags, image } = req.body
-
-            // Method 2:
+            const author = req.user._id
             const body = {
                 title: req.body.title || '',
-                content: req.body.content || 'Does not have content',
+                content: req.body.content || 'No content',
                 tags: req.body.tags || [],
-                image: req.body.image || '',
+                image: req.body.image || null,
             }
-
-            // Method 3:
-            // const fields = ['title', 'content', 'tags', 'image']
-            // const body = {}
-            // fields.forEach(field => {
-            //     body[field] = req.body[field]
-            // })
 
             // Create dreams in database
             const dreamServices = new DreamServices()
-            const dream = await dreamServices.createDream(
-                body.title,
-                undefined,
-                body.content,
-                body.tags,
-                body.image
-            )
+            const dream = await dreamServices.createDream({ ...body, author })
 
             // Return response
             res.send(new Response({ entity: { dream } }))
@@ -60,14 +48,25 @@ module.exports = {
     },
 
     deleteDream: async (req, res) => {
-        // TODO:
         const id = req.params.dream_id
         try {
             const dreamServices = new DreamServices()
-            const removalResult = await dreamServices.deleteDream(id)
+            const targetDream = await dreamServices.getDream(id)
 
-            // return response
-            res.send(new Response({ entity: { removalResult } }))
+            if (
+                String(targetDream.author) === req.user._id ||
+                req.user.isAdmin
+            ) {
+                const removedDream = await dreamServices.deleteDream(id)
+                if (removedDream) {
+                    res.send(new Response({ entity: { dream: removedDream } }))
+                } else {
+                    res.status(404).send(response.NOT_FOUND)
+                }
+            } else {
+                // Not allowed to remove other's dream except for admin
+                res.status(403).send(response.FORBIDDEN)
+            }
         } catch (err) {
             console.error(err)
             res.status(500).send(new Response({ msg: err, status: 500 }))
@@ -75,14 +74,16 @@ module.exports = {
     },
 
     getDream: async (req, res) => {
-        // TODO:
         const id = req.params.dream_id
         try {
             const dreamServices = new DreamServices()
             const dream = await dreamServices.getDream(id)
 
-            // return response
-            res.send(new Response({ entity: { dream } }))
+            if (dream) {
+                res.send(new Response({ entity: { dream } }))
+            } else {
+                res.status(404).send(response.NOT_FOUND)
+            }
         } catch (err) {
             console.error(err)
             res.status(500).send(new Response({ msg: err, status: 500 }))
@@ -91,19 +92,42 @@ module.exports = {
 
     updateDream: async (req, res) => {
         // FIXME: If a field is not provided, you will overwrite the original content
+
         const id = req.params.dream_id
+
         const updates = {
-            title: req.body.title || '',
-            content: req.body.content || 'Does not have content',
-            tags: req.body.tags || [],
-            image: req.body.image || '',
+            title: req.body.title,
+            date: req.body.date,
+            content: req.body.content,
+            tags: req.body.tags,
+            image: req.body.image,
         }
+        // Remove unprovided fields
+        Object.keys(updates).forEach(
+            (key) => updates[key] === undefined && delete updates[key]
+        )
+
         try {
             const dreamServices = new DreamServices()
-            const updatedDream = await dreamServices.updateDream(id, updates)
+            const targetDream = await dreamServices.getDream(id)
 
-            // return response
-            res.send(new Response({ entity: { updatedDream } }))
+            if (
+                String(targetDream.author) === req.user._id ||
+                req.user.isAdmin
+            ) {
+                const updatedDream = await dreamServices.updateDream(
+                    id,
+                    updates
+                )
+
+                if (updatedDream) {
+                    res.send(new Response({ entity: { dream: updatedDream } }))
+                } else {
+                    res.status(404).send(response.NOT_FOUND)
+                }
+            } else {
+                res.status(403).send(response.FORBIDDEN)
+            }
         } catch (err) {
             console.error(err)
             res.status(500).send(new Response({ msg: err, status: 500 }))
